Reduce mp_chattime to 10
The time when the game is over and getting ready to switch map, the scoreboard shows for way too long, making the server feel dormant. The match over music even finishes yet players are still waiting at the scoreboard screen waiting for the map to change for several seconds in silence. A small QoL fix to reduce this time to, at most, 10 seconds.