Reserve slots for all players on map change
This gives all players connected after a match ends a configurable grace
period (60 seconds by default) to join the next map. This ensures they
will not be kicked in case the server is temporarily over capacity due
to a person with a reserved slot being connected, or by being "ninja'd"
out of their slot.

Configure by changing the `sm_reserve_grace` convar. Set to `0` to
disable this feature.

This overrides the default `reservedslots.sp` plugin.